LPD-53397 move FriendlyURLSeparatorConfigurationManagerTemporarySwapper
