+++ modules/apps/friendly-url/friendly-url-test-util/src/main/java/com/liferay/friendly/url/configuration/manager/test/util/FriendlyURLSeparatorConfigurationManagerTemporarySwapper.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.friendly.url.test.util;
+package com.liferay.friendly.url.configuration.manager.test.util;
 
 import com.liferay.friendly.url.configuration.manager.FriendlyURLSeparatorConfigurationManager;
 import com.liferay.portal.kernel.exception.PortalException;
