+++ modules/apps/friendly-url/friendly-url-test-util/src/main/resources/com/liferay/friendly/url/test/util/packageinfo
@@ -1,1 +1,1 @@
-version 1.1.0
+version 1.0.0
