+++ modules/apps/layout/layout-page-template-test/src/testIntegration/java/com/liferay/layout/page/template/internal/info/item/provider/test/DisplayPageInfoItemFieldSetProviderTest.java
@@ -10,7 +10,7 @@ import com.liferay.asset.display.page.constants.AssetDisplayPageConstants;
 import com.liferay.asset.display.page.model.AssetDisplayPageEntry;
 import com.liferay.asset.display.page.portlet.AssetDisplayPageFriendlyURLProvider;
 import com.liferay.asset.display.page.service.AssetDisplayPageEntryLocalService;
-import com.liferay.friendly.url.test.util.FriendlyURLSeparatorConfigurationManagerTemporarySwapper;
+import com.liferay.friendly.url.configuration.manager.test.util.FriendlyURLSeparatorConfigurationManagerTemporarySwapper;
 import com.liferay.info.field.InfoField;
 import com.liferay.info.field.InfoFieldSet;
 import com.liferay.info.field.InfoFieldValue;
