+++ modules/apps/layout/layout-test/src/testIntegration/java/com/liferay/layout/portlet/test/LayoutFriendlyURLSeparatorCompositeTest.java
@@ -6,7 +6,7 @@
 package com.liferay.layout.portlet.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.friendly.url.test.util.FriendlyURLSeparatorConfigurationManagerTemporarySwapper;
+import com.liferay.friendly.url.configuration.manager.test.util.FriendlyURLSeparatorConfigurationManagerTemporarySwapper;
 import com.liferay.journal.constants.JournalFolderConstants;
 import com.liferay.journal.model.JournalArticle;
 import com.liferay.journal.test.util.JournalTestUtil;
